+++ targets/TARGET_NUVOTON/TARGET_M480/can_api.c
@@ -57,7 +57,7 @@ void can_init_freq(can_t *obj, PinName rd, PinName td, int hz)
 
     const struct nu_modinit_s *modinit = get_modinit(obj->can, can_modinit_tab);
     MBED_ASSERT(modinit != NULL);
-    MBED_ASSERT(modinit->modname == obj->can);
+    MBED_ASSERT(modinit->modname == (int) obj->can);
 
     // Reset this module
     SYS_ResetModule(modinit->rsetidx);
@@ -97,7 +97,7 @@ void can_free(can_t *obj)
     const struct nu_modinit_s *modinit = get_modinit(obj->can, can_modinit_tab);
 
     MBED_ASSERT(modinit != NULL);
-    MBED_ASSERT(modinit->modname == obj->can);
+    MBED_ASSERT(modinit->modname == (int) obj->can);
 
     // Reset this module
     SYS_ResetModule(modinit->rsetidx);
@@ -328,7 +328,7 @@ void can_reset(can_t *obj)
     const struct nu_modinit_s *modinit = get_modinit(obj->can, can_modinit_tab);
 
     MBED_ASSERT(modinit != NULL);
-    MBED_ASSERT(modinit->modname == obj->can);
+    MBED_ASSERT(modinit->modname == (int) obj->can);
 
     // Reset this module
     SYS_ResetModule(modinit->rsetidx);
